+++ beater/v1_handler.go
@@ -1,0 +1,17 @@
+package beater
+
+import (
+	"net/http"
+
+	"github.com/elastic/apm-server/decoder"
+	"github.com/elastic/apm-server/processor"
+)
+
+func (v *v1Route) Handler(p processor.Processor, beaterConfig *Config, report reporter) http.Handler {
+	decoder := v.configurableDecoder(beaterConfig, decoder.DecodeLimitJSONData(beaterConfig.MaxUnzippedSize))
+	tconfig := v.transformConfig(beaterConfig)
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		res := processRequest(r, p, tconfig, report, decoder)
+		sendStatus(w, r, res)
+	})
+}
